matrix finder now works on module instead of sensor 1.
doesn't really matter, was just a test, but this way I'm fixing a bug waiting to happen. the ideal overlap infos should not contain any matrices in the future. this code now works with idealDetectorMatrices.json as the sole source of matrices.

either way, the resultant overlap matrix is in PND global, "as all things should be".